Organized the classes into suitable java packages, Command, Task, etc
