--- src/main/java/Commands.java
@@ -1,10 +1,0 @@
-public enum Commands {
-    todo,
-    deadline,
-    mark,
-    unmark,
-    list,
-    bye,
-    delete,
-    event;
-}
