+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,4 +1,11 @@
-public class DeadlineCommand extends Command{
+package duke.command;
+
+import duke.task.Deadline;
+import duke.storage.StorageList;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class DeadlineCommand extends Command {
     private String message;
     private String timing;
