+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,13 +1,19 @@
-public class UnmarkCommand extends Command {
+package duke.command;
+
+import duke.storage.StorageList;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class DeleteCommand extends Command {
     private int listNum;
 
-    public UnmarkCommand(String input) {
+    public DeleteCommand(String input) {
         int tasknumbermark = Integer.valueOf(input) - 1;
         this.listNum = tasknumbermark;
     }
 
     public boolean execute(TaskList tasks, Ui ui, StorageList storage) {
-        tasks.unmarkTask(listNum);
+        tasks.deleteTask(listNum);
         return true;
     }
 }
