+++ src/main/java/duke/command/ListCommand.java
@@ -1,6 +1,13 @@
+package duke.command;
+
+import duke.storage.StorageList;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 import java.util.ArrayList;
 
-public class ListCommand extends Command{
+public class ListCommand extends Command {
     private ArrayList<Task> list;
 
     public boolean execute(TaskList tasks, Ui ui, StorageList storage) {
