+++ src/main/java/duke/command/MarkCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.storage.StorageList;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class MarkCommand extends Command {
     private int listNum;
