+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,7 @@
+package duke.parser;
+
+import duke.command.*;
+
 public class Parser {
     private enum Type {
         todo,
