+++ src/main/java/duke/storage/StorageList.java
@@ -1,9 +1,12 @@
+package duke.storage;
+
+import exception.DukeException;
+import duke.task.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Scanner;
